+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJQ4GJ40DCH2VJ3TPPCXWX2K"
+        "01KJQEWDQ5M989094QZZP9VQ6A"
     )
     r"""The unique identifier of the data source"""