+++ packages/orq-rc/RELEASES.md
@@ -6249,3 +6249,13 @@ Based on:
 - [python v4.5.0-rc.19] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.19] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.19 - packages/orq-rc
+
+## 2026-03-02 14:25:25
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.736.3 (2.845.15) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.21] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.21] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.21 - packages/orq-rc